Оптимизация прямой задачи

+++ directProblemQuadratureSamplingMethod/sourceFunction.h
@@ -1,12 +1,12 @@
 ﻿#pragma once
 #include "basicFunctions.h"
 
-// координаты источников, и фунция, характеризующая источник
+// координаты источников, и функция, характеризующая источник
 
-const Posize_t source01 = { -1.0f, 0.0f };
-const Posize_t source02 = { -1.0f, 2.5f };
-const Posize_t source03 = { -1.0f, 5.0f };
-const Posize_t source04 = { -1.0f, 7.5f };
-const Posize_t source05 = { -1.0f, 10.0f };
+const Posize_t source01 = { -1.0, 0.0 };
+const Posize_t source02 = { -1.0, 2.5 };
+const Posize_t source03 = { -1.0, 5.0 };
+const Posize_t source04 = { -1.0, 7.5 };
+const Posize_t source05 = { -1.0, 10.0 };
 
-std::complex<float> SourceFunction(const Posize_t source, const float x, const float y);
+std::complex<double> SourceFunction(const Posize_t source, const double x, const double y);
